requirements.txt
Flask
SQLAlchemy
Flask-SQLAlchemy
Flask-Migrate
ccy
alembic
psycopg2-binary
flask-mongoengine
pillow